Added the sender processing code and Receiving processing
The program processes the hammingcode for 2 bytes(16 bits). This hammingcode project has two parts. One is sender and the second is receiver. This is because different number of inputs are required in each. To process the data, enter input in sender code. The same processed and encoded output from sender can be checked through receiver program.